Simpler management for number of individuals

+++ kernel/evolution_selection.cc
@@ -41,8 +41,8 @@ namespace vita
   {
     const population &pop(evo_->population());
 
-    const unsigned n(pop.size());
-    const unsigned mate_zone(*pop.env().mate_zone);
+    const size_t n(pop.size());
+    const size_t mate_zone(*pop.env().mate_zone);
     const unsigned rounds(*pop.env().tournament_size);
 
     index_t sel(random::ring(target, mate_zone, n));
@@ -73,8 +73,8 @@ namespace vita
   {
     const population &pop(evo_->population());
 
-    const unsigned n(pop.size());
-    const unsigned mate_zone(*pop.env().mate_zone);
+    const size_t n(pop.size());
+    const size_t mate_zone(*pop.env().mate_zone);
     const unsigned rounds(*pop.env().tournament_size);
     const index_t target(random::between<index_t>(0, n));
 
@@ -90,9 +90,11 @@ namespace vita
       const fitness_t new_fitness(evo_->fitness(pop[new_index]));
 
       index_t j(0);
+      // Where is the insertion point?
       while (j < i && new_fitness < evo_->fitness(pop[ret[j]]))
         ++j;
 
+      // Shift right elements after the insertion point.
       for (index_t k(j); k < i; ++k)
         ret[k + 1] = ret[k];
 
@@ -124,9 +126,9 @@ namespace vita
   {
     const population &pop(evo_->population());
 
-    const unsigned n(pop.size());
-    const unsigned mate_zone(*pop.env().mate_zone);
-    const unsigned size(*pop.env().tournament_size);
+    const size_t n(pop.size());
+    const size_t mate_zone(*pop.env().mate_zone);
+    const size_t size(*pop.env().tournament_size);
     const index_t target(random::between<index_t>(0, n));
 
     assert(size);
